IP: PWM_2.0: Add Optional GPIO Interface
In order to allow RGB LED interfaces in board files to be connected to this IP, I have added a gpio_rtl interface which can be enabled or disabled. When enabled, the standard PWM data_rtl interface is disabled, and vice versa. The GPIO interface is disabled by default, to maintain backwards compatibility.