IP: PWM_2.0: Board File Integration
Modified the ip so that it can connect to non-input-only gpio interfaces in Digilent board files. This is primarily intended to work with LEDs and RGB LEDs, and has been tested with only the Cmod S7 so far. In order to make the interface work properly, each gpio interface in the board files should be updated to include a PWM preset. The only preset parameter that needs to be included is CONFIG.POLARITY.