IP: Pmods: PmodESP32: Initial Commit
TODO:
The AXI UartLite IP used should be replaced by an AXI UARTNS550 IP, in order to allow AXI clock frequencies other than 100MHz.
The AT command "ATE0" should be sent to the ESP32 before control of the command line interface is handed over to the user, in order to prevent double prints.